6.5 - 6.8

+++ DesignPatterns/src/main/kotlin/dogs/Dog.kt
@@ -9,7 +9,7 @@ data class Dog(
     @SerialName("dog_breed")
     val dogBreed: String,
     @SerialName("dog_id")
-    val dogId: Long,
+    val dogId: Int,
     @SerialName("dog_name")
     val dogName: String,
     @SerialName("dog_weight")
